requirements.txt
aiofiles==0.7.0
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.5
click==8.0.1
fastapi==0.68.1
h11==0.12.0
httptools==0.2.0
idna==3.2
pydantic==1.8.2
python-dotenv==0.19.0
PyYAML==5.4.1
requests==2.26.0
rfeed==1.1.1
selenium==3.141.0
starlette==0.14.2
typing-extensions==3.10.0.2
urllib3==1.26.6
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==10.0
youtube-dl==2021.6.6
